backend/requirement.txt
paho-mqtt==1.5.1
python-dotenv==0.21.0
